Add router all Science Journal
| `:journal` |    Full Name of the Journal    | Route                                                            |
| :--------: | :----------------------------: | ---------------------------------------------------------------- |
|  science   |            Science             | [/sciencemag/current/science](/sciencemag/current/science)       |
|  advances  |        Science Advances        | [/sciencemag/current/advances](/sciencemag/current/advances)     |
| immunology |       Science Immunology       | [/sciencemag/current/immunology](/sciencemag/current/immunology) |
|  robotics  |        Science Robotics        | [/sciencemag/current/robotics](/sciencemag/current/robotics)     |
|    stke    |       Science Signaling        | [/sciencemag/current/stke](/sciencemag/current/stke)             |
|    stm     | Science Translational Medicine | [/sciencemag/current/stm](/sciencemag/current/stm)               |
