+++ lib/router.js
@@ -2014,10 +2014,8 @@ router.get('/nature/news', require('./routes/nature/news'));
 router.get('/nature/highlight', require('./routes/nature/highlight'));
 
 // science [Sci Journal]
-router.get('/sciencemag/science/current', require('./routes/sciencemag/science/current'));
-router.get('/sciencemag/science/early', require('./routes/sciencemag/science/early'));
-// science 子刊 [Sci Journal]
-router.get('/sciencemag/advances/current', require('./routes/sciencemag/advances/current'));
+router.get('/sciencemag/current/:journal?', require('./routes/sciencemag/current'));
+router.get('/sciencemag/early/science', require('./routes/sciencemag/early'));
 
 // dlsite
 router.get('/dlsite/new/:type', require('./routes/dlsite/new'));
