+++ lib/routes/sciencemag/early.js
@@ -1,25 +1,15 @@
+// only support Science journal
+
 const cheerio = require('cheerio');
 const got = require('@/utils/got');
 
 module.exports = async (ctx) => {
-    const base = `https://science.sciencemag.org`;
+    const base = `https://science.sciencemag.org/content/early/recent`;
 
     const res = await got.get(base);
     const pageCapture = cheerio.load(res.data);
 
-    // just select paper relative sections
-    const sectionList = ['issue-toc-section-research-articles', 'issue-toc-section-review', 'issue-toc-section-reports'];
-    const list = [].concat.apply(
-        [],
-        sectionList.map((section) => {
-            const sec = cheerio.load(pageCapture(`ul > li .issue-toc-section.${section}`).html());
-            const sectionName = sec('h2').text();
-            const sectionList = sec('ul > li > div > div > article > div')
-                .append(`<div class="toc-section-type">${sectionName}</div>`)
-                .get();
-            return sectionList;
-        })
-    );
+    const list = pageCapture('ul > li > div > article > div').get();
 
     const out = await Promise.all(
         list.map(async (item) => {
@@ -28,7 +18,6 @@ module.exports = async (ctx) => {
             const title = $('h3').text();
             const partial = $('h3 > a').attr('href');
             const address = `${base}${partial}`;
-            const section = $('div .toc-section-type').text();
 
             let author;
             const authorList = $('span.highwire-citation-authors > span.highwire-citation-author')
@@ -50,15 +39,6 @@ module.exports = async (ctx) => {
             }
 
             // contents
-            // section content
-            let sectionContents = '';
-            if (section !== '') {
-                sectionContents = `
-                <div id="content-section">
-                  <span style="color: #d40016; text-transform: uppercase; font-weight: 700;">${section}</span>
-                </div>
-                `;
-            }
             // brief content
             const brief = $('div.highwire-cite-snippet > div > div > p').text();
             let briefContents = '';
@@ -69,16 +49,27 @@ module.exports = async (ctx) => {
                 </div>
                 `;
             }
+
             const itemPage = await got.get(address);
             const itemCapture = cheerio.load(itemPage.data);
+            // section content
+            const section = itemCapture('header > div.overline').text();
+            let sectionContents = '';
+            if (section !== '') {
+                sectionContents = `
+                <div id="content-section">
+                  <span style="color: #d40016; text-transform: uppercase; font-weight: 700;">${section}</span>
+                  <span style="color: grey; text-transform: uppercase; font-weight: 700;">[Published Online]</span>
+                </div>
+                `;
+            }
+            // abs content
             const abs = itemCapture('div > div.abstract-view > div.section')
                 .map(function(i, el) {
                     return $(el).html();
                 })
                 .get()
                 .join('<br>');
-
-            // abs content
             let absContents = '';
             if (abs !== null) {
                 absContents = `
@@ -102,8 +93,8 @@ module.exports = async (ctx) => {
         })
     );
     ctx.state.data = {
-        title: `Science | Current Table of Contents`,
-        description: `Science, a research journal`,
+        title: `Science | First Release`,
+        description: `Science, a research journal. For papers that published online.`,
         link: base,
         item: out,
     };
